requirements.txt
# Core API server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LLM & RAG
langchain==0.1.0
langchain_community==0.0.10
langchain-text-splitters==0.0.1
langchain_chroma==0.1.0
unstructured==0.21.5
openai==1.3.0

# ✓ CONSTRAINT: Embedding Model (all-MiniLM-L6-v2)
sentence-transformers==2.7.0
torch==2.2.0

# ✓ Vector Database - ChromaDB (No Docker required!)
chromadb==0.4.22

# Vector math (cosine similarity)
numpy==1.24.3
scipy==1.11.0
scikit-learn==1.3.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Additional
python-multipart==0.0.6

